feat(BlueMind4): based on original BlueMind plugin
not certain adding a plugin specific to bm4 is the best way to do this.
on the other hand, I'm not sure we can check which bluemind version is in
use, in the plugin constructor. And I'ld rather not have useless checks
catching ENOENTs.

as a sidenote, there also is a bm-sds-proxy service that we might want
to monitor. I haven't used nor configured it so far, all I can see from
my test setups are JVM memory metrics and other generic counters,
nothing specific to SDS, ... left as a TODO.