fix(BlueMind): introduce BlueMind4 support
not certain overloading the existing bluemind plugin is the best way to
do this. Assuming modes could be individually enabled/disabled, then
it avoids duplicating code.
-- with BlueMind3, the mapi mode should not be used.
-- with BlueMind4, the ips mode should not be used.

as a sidenote, there also is a bm-sds-proxy service that we might want
to monitor. I haven't used nor configured it so far, all I can see from
my test setups are JVM memory metrics and other generic counters,
nothing specific to SDS, ... left as a TODO.